+++ lib/n_plus_one_control/rspec.rb
@@ -4,7 +4,7 @@ gem "rspec-core", ">= 3.5"
 
 require "n_plus_one_control"
 require "n_plus_one_control/rspec/dsl"
-require "n_plus_one_control/rspec/matcher"
+require "n_plus_one_control/rspec/matchers/perform_constant_number_of_queries"
 require "n_plus_one_control/rspec/context"
 
 module NPlusOneControl
